+++ mmdet/core/bbox/samplers/random_sampler.py
@@ -3,7 +3,7 @@ import torch
 
 from ..builder import BBOX_SAMPLERS
 from .base_sampler import BaseSampler
-
+from utils import int_dtype
 
 @BBOX_SAMPLERS.register_module()
 class RandomSampler(BaseSampler):
@@ -51,11 +51,11 @@ class RandomSampler(BaseSampler):
                 device = torch.cuda.current_device()
             else:
                 device = 'cpu'
-            gallery = torch.tensor(gallery, dtype=torch.long, device=device)
+            gallery = torch.tensor(gallery, dtype=int_dtype, device=device)
         # This is a temporary fix. We can revert the following code
         # when PyTorch fixes the abnormal return of torch.randperm.
         # See: https://github.com/open-mmlab/mmdetection/pull/5014
-        perm = torch.randperm(gallery.numel())[:num].to(device=gallery.device)
+        perm = torch.randperm(gallery.numel(), dtype=int_dtype)[:num].to(device=gallery.device)
         rand_inds = gallery[perm]
         if not is_tensor:
             rand_inds = rand_inds.cpu().numpy()
